dont link with libgcc
if the libgcc was compiled without certain flags, such as
-mno-check-zero-division, it may contain instruction that assume that
you are running in a managed environment (usermode).

for example, on mips, i get trap instructions, which doesn't make much
sense for freestanding code.

so, i now provide my own implementation of libgcc functions, which is
compiled using my own cflags, which ensure that none of that bullshit
will appear in the compiler's output.